# installation: make install output more useful