- [x] We used to print out where source code was staged. Bring this back.
- [x] Simplify message when binaries are not found (we don't need the name/hash)
- [x] Clean up warning when modules are not regenerated. Users are distracted by this; they think there is an issue. Make it a simple message with instructions to regenerate.

Before:

```
==> Installing zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
==> No binary for zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach found: installing from source
==> Using cached archive: /Users/gamblin2/Workspace/spack/var/spack/cache/_source-cache/archive/c3/c3e5e9fdd5004dcb542feda5ee4f0ff0744628baf8ed2dd5d66f8ca1197cb1a1.tar.gz
==> No patches needed for zlib
==> zlib: Executing phase: 'install'
==> Warning: Module file already exists : skipping creation
file : /Users/gamblin2/Workspace/spack/share/spack/modules/darwin-catalina-skylake/zlib-1.2.11-apple-clang-12.0.0-f7owbzf
spec : zlib@1.2.11%apple-clang@12.0.0+optimize+pic+shared arch=darwin-catalina-skylake
==> zlib: Successfully installed zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
  Fetch: 0.00s.  Build: 3.65s.  Total: 3.66s.
[+] /Users/gamblin2/Workspace/spack/opt/spack/darwin-catalina-skylake/apple-clang-12.0.0/zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
```

After:

```
==> Installing zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
==> No binary found, installing from source
==> Using cached archive: /Users/gamblin2/Workspace/spack/var/spack/cache/_source-cache/archive/c3/c3e5e9fdd5004dcb542feda5ee4f0ff0744628baf8ed2dd5d66f8ca1197cb1a1.tar.gz
==> Source code staged at: /var/folders/0s/q_y0zhfj6xdd5n7rn780zz6h001qr7/T/gamblin2/spack-stage/spack-stage-zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
==> No patches needed for zlib
==> zlib: Executing phase: 'install'
==> Module file exists; won't overwrite. Use `spack module` to refresh.
==> zlib: Successfully installed zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
  Fetch: 0.00s.  Build: 3.22s.  Total: 3.22s.
[+] /Users/gamblin2/Workspace/spack/opt/spack/darwin-catalina-skylake/apple-clang-12.0.0/zlib-1.2.11-f7owbzfu4zsr6yn7x7hljransoof6ach
```