@mwkrentel 2021-05-16
@tgamblin  What I'd really like `spack install` to display is a fuller spec.
Not all the dependencies, that would be too much.  But could we add variants?